fixed size_t overflow issue affecting number of hash buckets

+++ src/WordFormablePowerInts.h
@@ -15,6 +15,6 @@
 #include <stdio.h>
 #include <string.h>
 
-void processTokensFromFile(char* base_str, FILE* input_file, char* c_buff, const int max_length, const int silence, const int buckets);
+void processTokensFromFile(char* base_str, FILE* input_file, char* c_buff, const unsigned int max_length, const int unsigned silence, const size_t buckets);
 
 #endif
